working on the implementation of the hierarchical OOD implementation

+++ Contrastive_uncertainty/general_hierarchy/train/train_general_hierarchy.py
@@ -47,7 +47,7 @@ def train(params,model_module,model_function):
 
     wandb_logger.watch(model, log='gradients', log_freq=100) # logs the gradients
 
-    trainer = pl.Trainer(fast_dev_run = config['fast_run'],progress_bar_refresh_rate=20,
+    trainer = pl.Trainer(precision =16, fast_dev_run = config['fast_run'],progress_bar_refresh_rate=20,
                         limit_train_batches = config['training_ratio'],limit_val_batches=config['validation_ratio'],limit_test_batches = config['test_ratio'],
                         max_epochs = config['epochs'],check_val_every_n_epoch = config['val_check'],
                         gpus=1,logger=wandb_logger,checkpoint_callback = False,deterministic =True,callbacks = desired_callbacks)#,auto_lr_find = True)
